refactor(sensors): Change Angle to analog
