+++ src/main/java/frc/robot/sensors/analogsensor/AnalogSensor.java
@@ -1,0 +1,8 @@
+package frc.robot.sensors.analogsensor;
+
+public abstract class AnalogSensor {
+
+  public abstract double getAngle();
+
+  public abstract void reset();
+}
