--- src/main/java/frc/robot/sensors/anglesensor/AngleSensor.java
@@ -1,8 +1,0 @@
-package frc.robot.sensors.anglesensor;
-
-public abstract class AngleSensor {
-
-  public abstract double getAngle();
-
-  public abstract void reset();
-}
